fix int to float error

+++ SpyTeam3030_Prototye/Assets/Scripts/AI/CombatController.cs
@@ -24,7 +24,7 @@ public class CombatController : NetworkBehaviour
 	[SyncVar]
 	public int cardID;
     [SyncVar]
-    public int restoreHealth;
+    public float restoreHealth;
 
     [Header("Combat Display")] 
     public GameObject healthBar;
